% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPeaks.R
\name{alignPeaks}
\alias{alignPeaks}
\title{Simplest possible feature detection method}
\usage{
alignPeaks(peaks, epsilon = 0.005)
}
\arguments{
\item{peaks}{data.frame with columns 'mz' and 'time' for m/z values and scan ids.}

\item{epsilon}{width of feature allowed region}
}
\value{
peaks data.frame with column 'ion', which contains feature m/z value
for the feature peaks and NA otherwise
}
\description{
Gets the highest peak in the dataset and collect all other peaks in its
vicinity. Finally filter multiple peaks in one scan by choosing the highest one.
}
\seealso{
Other feature detection functions: \code{\link{alignPeaksDTppm}},
  \code{\link{alignPeaksDTrle}}, \code{\link{alignPeaksDT}}
}
