% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPeaks.R
\name{alignPeaksDT}
\alias{alignPeaksDT}
\title{Feature detection by linear regression}
\usage{
alignPeaksDT(peaks, epsilon = 0.005, minAbsLen = 300, minRelLen = 0.5,
  minRunLen = 100, thRes = 2)
}
\arguments{
\item{peaks}{data.frame with columns \code{mz} and \code{time} for m/z values and scan ids.}

\item{epsilon}{width of feature allowed region. Default 5e-3.}

\item{minAbsLen}{minimal length for the continious feature in number of scans.
Default 300.}

\item{minRelLen}{minimal length for the continious feature relative to the
total length of the registration. Default 5e-1 (50\%).}

\item{minRunLen}{minimal length of uninterapted nonzero run for the
continious feature. Default 100.}

\item{thRes}{threshold for standartized linear regression value. Default 2.0.}
}
\value{
peaks data.frame with column \code{ion}, which contains feature m/z value
for the feature peaks and NA otherwise
}
\description{
Check the precondition for the continious feature and if met fit linear
regression to the data and assign to the feature all peaks for which
standartized residual is less then \code{\link{thRes}}.
}
\details{
The preconditions are:
\itemise
}
\seealso{
Other feature detection functions: \code{\link{alignPeaksDTppm}},
  \code{\link{alignPeaksDTrle}}, \code{\link{alignPeaks}}
}
