% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPeaks.R
\name{alignPeaksDTrle}
\alias{alignPeaksDTrle}
\title{Feature detection by clusterisation and linear regression}
\usage{
alignPeaksDTrle(peaks, ppm = 20, minAbsLen = 300, minRelLen = 0.5,
  minRunLen = 100, thRes = 2)
}
\arguments{
\item{peaks}{data.frame with columns 'mz' for the m/z values,
'time' for the scan time value and
'scan' for the scan ids.}

\item{ppm}{width of feature allowed region}

\item{minAbsLen}{minimal length for the continious feature in number of scans.
See also alignPeaksDT.}

\item{minRelLen}{minimal length for the continious feature relative to the
total length of the registration. See also alignPeaksDT.}

\item{minRunLen}{minimal length of uninterapted nonzero run for the
continious feature. See also alignPeaksDT.}

\item{thRes}{threshold for standartized linear regression value. See also alignPeaksDT.}
}
\value{
peaks data.frame with column 'ion', which contains feature m/z value
for the feature peaks and NA otherwise
}
\description{
Feature detection by clusterisation and linear regression
}
\seealso{
Other feature detection functions: \code{\link{alignPeaksDTppm}},
  \code{\link{alignPeaksDT}}, \code{\link{alignPeaks}}
}
