% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutTreeHybrid.R
\name{cutTreeHybrid}
\alias{cutTreeHybrid}
\title{Get clusters in m/z domain}
\usage{
cutTreeHybrid(p.)
}
\arguments{
\item{p.}{data.frame with column 'mz' containing m/z values to cluster}
}
\value{
result of cutreeDynamic call
}
\description{
Function calculates hierarchical clustering with 'average' method by hclust.
The number of clusters and their composition is detected by cutreeDynamic function.
}
\examples{

p<-data.frame(mz=c(1.1*runif(100),1.5*runif(100),3.1*runif(100),5.1*runif(100),))
p$time<-seq_along(p$mz)
cth<-cutTreeHybrid(p)
.pal <- brewer.pal(12,'Paired')
p$color<-.pal[cth+1]
qplot(time,mz,data = p.,color=color)

}
