% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle.R
\name{get_Rle}
\alias{get_Rle}
\title{Calculate Rle representation of feature in time domain.}
\usage{
get_Rle(scan, length)
}
\arguments{
\item{scan}{vector indices of nonzero intensity values}

\item{length}{total number of scans in the spectrum}
}
\value{
Rle representation of the feature in time domain
}
\description{
Function creates vector of zeros and for each nonzero scan
sets the value to 1. The vector obtained converted into Rle
representation and returned
}
