% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle.R
\name{get_Rle_nogaps}
\alias{get_Rle_nogaps}
\title{Calculate Rle representation of feature in time domain with gaps.}
\usage{
get_Rle_nogaps(scan, length, gap = 1, factor = 10)
}
\arguments{
\item{scan}{vector indices of nonzero intensity values (see get_Rle)}

\item{length}{total number of scans in the spectrum (see get_Rle)}

\item{gap}{max length of gap between two nonzero runs}

\item{factor}{}
}
\value{
with gaps filled
}
\description{
Function takes Rle representation created by get_Rle and fill gaps
up to gap in length.
}
