% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle.R
\name{get_runLen}
\alias{get_runLen}
\title{Vector of length of feature nonzero runs}
\usage{
get_runLen(scan, length)
}
\arguments{
\item{scan}{vector indices of nonzero intensity values (see get_Rle)}

\item{length}{total number of scans in the spectrum (see get_Rle)}
}
\value{
vector of length of continious nonzero runs
}
\description{
Vector of length of feature nonzero runs
}
