\name{lmFeatures-package}
\alias{lmFeatures-package}
\alias{lmFeatures}
\docType{package}
\title{
\packageTitle{lmFeatures}
}
\description{
\packageDescription{lmFeatures}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lmFeatures}
\packageIndices{lmFeatures}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lmFeatures}

Maintainer: \packageMaintainer{lmFeatures}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
