% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePeaks.R
\name{makePeaks}
\alias{makePeaks}
\title{Converts XCMS data into MZ table}
\usage{
makePeaks(xcms)
}
\arguments{
\item{xcms}{xcmzRaw data to convert}
}
\value{
ata.table with five columns: 'mz', 'intensity', 'ion', 'time', 'scan'
}
\description{
Creates data.table with five columns: 'mz', 'intensity', 'ion', 'time', 'scan'
}
