% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Custom print table function}
\usage{
print_table(mat, digits = -5)
}
\arguments{
\item{mat}{data to be printed matrix, or data convertable to matrix.}

\item{digits}{cutoff and rounding parameter. See xtable documentation.}
}
\value{
none
}
\description{
Set up defaults for printing the LaTeX table with xtable. Function do not
return anything it just sends strings to the output. Useful for use with
knitr or Sweave.
}
