% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_tableL}
\alias{print_tableL}
\title{Print table in the landscape mode.}
\usage{
print_tableL(mat, digits = -5)
}
\arguments{
\item{mat}{data to be printed matrix, or data convertable to matrix.}

\item{digits}{cutoff and rounding parameter. See xtable documentation.}
}
\value{
none
}
\description{
See print_table.
}
